Support hunyuan offload and teacache.
