+++ lightx2v/__main__.py
@@ -1,5 +1,4 @@
 import argparse
-from contextlib import contextmanager
 import torch
 import torch.distributed as dist
 import os
@@ -9,6 +8,7 @@ import json
 import torchvision
 import torchvision.transforms.functional as TF
 import numpy as np
+from contextlib import contextmanager
 from PIL import Image
 from lightx2v.text2v.models.text_encoders.hf.llama.model import TextEncoderHFLlamaModel
 from lightx2v.text2v.models.text_encoders.hf.clip.model import TextEncoderHFClipModel
@@ -16,14 +16,13 @@ from lightx2v.text2v.models.text_encoders.hf.t5.model import T5EncoderModel
 from lightx2v.text2v.models.text_encoders.hf.llava.model import TextEncoderHFLlavaModel
 
 from lightx2v.text2v.models.schedulers.hunyuan.scheduler import HunyuanScheduler
-from lightx2v.text2v.models.schedulers.hunyuan.feature_caching.scheduler import HunyuanSchedulerFeatureCaching
+from lightx2v.text2v.models.schedulers.hunyuan.feature_caching.scheduler import HunyuanSchedulerTaylorCaching, HunyuanSchedulerTeaCaching
 from lightx2v.text2v.models.schedulers.wan.scheduler import WanScheduler
-from lightx2v.text2v.models.schedulers.wan.feature_caching.scheduler import WanSchedulerFeatureCaching
+from lightx2v.text2v.models.schedulers.wan.feature_caching.scheduler import WanSchedulerTeaCaching
 
 from lightx2v.text2v.models.networks.hunyuan.model import HunyuanModel
 from lightx2v.text2v.models.networks.wan.model import WanModel
 from lightx2v.text2v.models.networks.wan.lora_adapter import WanLoraWrapper
-
 
 from lightx2v.text2v.models.video_encoders.hf.autoencoder_kl_causal_3d.model import VideoEncoderKLCausal3DModel
 from lightx2v.text2v.models.video_encoders.hf.wan.vae import WanVAE
@@ -38,6 +37,7 @@ def time_duration(label: str = ""):
     yield
     end_time = time.time()
     print(f"==> {label} start:{time.strftime('%Y-%m-%d %H:%M:%S', time.localtime(start_time))} cost {end_time - start_time:.2f} seconds")
+    torch.cuda.synchronize()
 
 
 def load_models(args, model_config):
@@ -61,16 +61,16 @@ def load_models(args, model_config):
         vae_model = VideoEncoderKLCausal3DModel(args.model_path, dtype=torch.float16, device=init_device, args=args)
 
     elif args.model_cls == "wan2.1":
-        text_encoder = T5EncoderModel(
-            text_len=model_config["text_len"],
-            dtype=torch.bfloat16,
-            device=init_device,
-            checkpoint_path=os.path.join(args.model_path, "models_t5_umt5-xxl-enc-bf16.pth"),
-            tokenizer_path=os.path.join(args.model_path, "google/umt5-xxl"),
-            shard_fn=None,
-        )
-        text_encoders = [text_encoder]
-
+        with time_duration("Load Text Encoder"):
+            text_encoder = T5EncoderModel(
+                text_len=model_config["text_len"],
+                dtype=torch.bfloat16,
+                device=init_device,
+                checkpoint_path=os.path.join(args.model_path, "models_t5_umt5-xxl-enc-bf16.pth"),
+                tokenizer_path=os.path.join(args.model_path, "google/umt5-xxl"),
+                shard_fn=None,
+            )
+            text_encoders = [text_encoder]
         with time_duration("Load Wan Model"):
             model = WanModel(args.model_path, model_config, init_device)
 
@@ -256,8 +256,10 @@ def init_scheduler(args, image_encoder_output):
     if args.model_cls == "hunyuan":
         if args.feature_caching == "NoCaching":
             scheduler = HunyuanScheduler(args, image_encoder_output)
+        elif args.feature_caching == "Tea":
+            scheduler = HunyuanSchedulerTeaCaching(args, image_encoder_output)
         elif args.feature_caching == "TaylorSeer":
-            scheduler = HunyuanSchedulerFeatureCaching(args, image_encoder_output)
+            scheduler = HunyuanSchedulerTaylorCaching(args, image_encoder_output)
         else:
             raise NotImplementedError(f"Unsupported feature_caching type: {args.feature_caching}")
 
@@ -265,7 +267,7 @@ def init_scheduler(args, image_encoder_output):
         if args.feature_caching == "NoCaching":
             scheduler = WanScheduler(args)
         elif args.feature_caching == "Tea":
-            scheduler = WanSchedulerFeatureCaching(args)
+            scheduler = WanSchedulerTeaCaching(args)
         else:
             raise NotImplementedError(f"Unsupported feature_caching type: {args.feature_caching}")
 
@@ -338,7 +340,6 @@ if __name__ == "__main__":
     parser.add_argument("--use_bfloat16", action="store_true", default=True)
     parser.add_argument("--lora_path", type=str, default=None)
     parser.add_argument("--strength_model", type=float, default=1.0)
-
     args = parser.parse_args()
 
     start_time = time.time()
@@ -383,7 +384,8 @@ if __name__ == "__main__":
     else:
         image_encoder_output = {"clip_encoder_out": None, "vae_encode_out": None}
 
-    text_encoder_output = run_text_encoder(args, args.prompt, text_encoders, model_config, image_encoder_output)
+    with time_duration("Run Text Encoder"):
+        text_encoder_output = run_text_encoder(args, args.prompt, text_encoders, model_config, image_encoder_output)
 
     set_target_shape(args, image_encoder_output)
     scheduler = init_scheduler(args, image_encoder_output)
@@ -399,16 +401,15 @@ if __name__ == "__main__":
         del text_encoder_output, image_encoder_output, model, text_encoders, scheduler
         torch.cuda.empty_cache()
 
-    images = run_vae(latents, generator, args)
+    with time_duration("Run VAE"):
+        images = run_vae(latents, generator, args)
 
     if not args.parallel_attn_type or (args.parallel_attn_type and dist.get_rank() == 0):
-        save_video_st = time.time()
-        if args.model_cls == "wan2.1":
-            cache_video(tensor=images, save_file=args.save_video_path, fps=16, nrow=1, normalize=True, value_range=(-1, 1))
-        else:
-            save_videos_grid(images, args.save_video_path, fps=24)
-        save_video_et = time.time()
-        print(f"Save video cost: {save_video_et - save_video_st}")
+        with time_duration("Save video"):
+            if args.model_cls == "wan2.1":
+                cache_video(tensor=images, save_file=args.save_video_path, fps=16, nrow=1, normalize=True, value_range=(-1, 1))
+            else:
+                save_videos_grid(images, args.save_video_path, fps=24)
 
     end_time = time.time()
     print(f"Total cost: {end_time - start_time}")
