+++ lightx2v/text2v/models/schedulers/scheduler.py
@@ -10,3 +10,6 @@ class BaseScheduler:
     def step_pre(self, step_index):
         self.step_index = step_index
         self.latents = self.latents.to(dtype=torch.bfloat16)
+
+    def clear(self):
+        pass
