+++ lightx2v/text2v/models/schedulers/wan/feature_caching/scheduler.py
@@ -2,7 +2,7 @@ import torch
 from ..scheduler import WanScheduler
 
 
-class WanSchedulerFeatureCaching(WanScheduler):
+class WanSchedulerTeaCaching(WanScheduler):
     def __init__(self, args):
         super().__init__(args)
         self.cnt = 0
