+++ lightx2v/text2v/models/schedulers/wan/scheduler.py
@@ -341,6 +341,3 @@ class WanScheduler(BaseScheduler):
             self.lower_order_nums += 1
 
         self.latents = prev_sample
-
-    def clear(self):
-        pass
